# openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Threads-like Application API",
		"description": "This is a Threads-like application backend API that allows users to set a username, post content, follow other users, and interact via comments and likes. No formal authentication required - usernames are stored in browser localStorage.",
		"version": "1.0.0",
		"contact": {
			"name": "API Support",
			"email": "support@example.com"
		}
	},
	"servers": [
		{
			"url": "https://api.example.com/v1",
			"description": "Production server"
		},
		{
			"url": "https://staging-api.example.com/v1",
			"description": "Staging server"
		}
	],
	"paths": {
...
		"/posts": {
			"get": {
				"tags": ["Posts"],
				"summary": "Get posts list",
				"description": "Retrieve a list of posts",
				"operationId": "get_posts_list_posts_get",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"description": "Page number for pagination (1-indexed)",
						"schema": {
							"type": "integer",
							"default": 1,
							"minimum": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Number of items per page",
						"schema": {
							"type": "integer",
							"default": 10,
							"minimum": 1,
							"maximum": 100
						}
					}
				],
				"responses": {
					"200": {
						"description": "Posts retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PostList" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Invalid or expired session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Posts"],
				"summary": "Create a new post",
				"description": "Add a new post to the authenticated user's profile",
				"operationId": "create_post_posts_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/PostCreate" }
						}
					},
					"required": true,
					"description": "Post content to create"
				},
				"responses": {
					"201": {
						"description": "Post created successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PostDetail" }
							}
						}
					},
					"400": {
						"description": "Bad Request - Invalid input data",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Invalid or expired session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/posts/{postId}": {
			"get": {
				"tags": ["Posts"],
				"summary": "Get post details by ID",
				"description": "Retrieve complete information about a specific post",
				"operationId": "get_post_detail_posts__postId__get",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to retrieve",
						"schema": { "type": "integer", "title": "Post ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": false,
						"description": "Username for authentication (optional)",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Post details retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PostDetail" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Posts"],
				"summary": "Delete a post",
				"description": "Remove a post created by the authenticated user",
				"operationId": "delete_post_posts__postId__delete",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to delete",
						"schema": { "type": "integer", "title": "Post ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": true,
						"description": "Username for authentication",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Post deleted successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/SuccessResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Not your post or invalid session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Posts"],
				"summary": "Update a post",
				"description": "Modify content of a post created by the authenticated user",
				"operationId": "update_post_posts__postId__put",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to update",
						"schema": { "type": "integer", "title": "Post ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": true,
						"description": "Username for authentication",
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/PostUpdate" }
						}
					},
					"description": "Updated post content"
				},
				"responses": {
					"200": {
						"description": "Post updated successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PostDetail" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Not your post or invalid session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/posts/{postId}/like": {
			"post": {
				"tags": ["Posts"],
				"summary": "Like a post",
				"description": "Add the authenticated user's like to a post",
				"operationId": "like_post_posts__postId__like_post",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to like",
						"schema": { "type": "integer", "title": "Post ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": true,
						"description": "Username for authentication",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Post liked successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/LikeResponse" }
							}
						}
					},
					"400": {
						"description": "Bad Request - Already liked this post",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Posts"],
				"summary": "Unlike a post",
				"description": "Remove the authenticated user's like from a post",
				"operationId": "unlike_post_posts__postId__like_delete",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to unlike",
						"schema": { "type": "integer", "title": "Post ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": true,
						"description": "Username for authentication",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Post unliked successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/LikeResponse" }
							}
						}
					},
					"400": {
						"description": "Bad Request - Haven't liked this post",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail",
						"description": "List of validation errors"
					}
				},
				"type": "object",
				"title": "HTTPValidationError",
				"description": "Error response for validation failures"
			},
			"PostCreate": {
				"properties": {
					"content": {
						"type": "string",
						"title": "Content",
						"description": "Text content of the post",
						"minLength": 1,
						"maxLength": 500
					},
					"username": {
						"type": "string",
						"title": "Username",
						"description": "Username of the post author (from localStorage)"
					}
				},
				"type": "object",
				"required": ["content", "username"],
				"title": "PostCreate"
			},
			"PostUpdate": {
				"properties": {
					"content": {
						"type": "string",
						"title": "Content",
						"description": "Updated text content of the post",
						"minLength": 1,
						"maxLength": 500
					}
				},
				"type": "object",
				"required": ["content"],
				"title": "PostUpdate"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location",
						"description": "Location of the validation error"
					},
					"msg": {
						"type": "string",
						"title": "Message",
						"description": "Validation error message"
					},
					"type": {
						"type": "string",
						"title": "Error Type",
						"description": "Type of validation error"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError",
				"description": "Single validation error details"
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string",
						"description": "Error message"
					},
					"status_code": {
						"type": "integer",
						"description": "HTTP status code"
					}
				},
				"required": ["error", "status_code"],
				"title": "ErrorResponse",
				"description": "Standard error response format"
			},
			"SuccessResponse": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string",
						"description": "Success message"
					}
				},
				"required": ["message"],
				"title": "SuccessResponse",
				"description": "Standard success response format"
			},
...
			"PostDetail": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"description": "Post ID"
					},
					"content": {
						"type": "string",
						"description": "Post content"
					},
					"author": {
						"type": "object",
						"properties": {
							"id": {
								"type": "integer",
								"description": "Author user ID"
							},
							"username": {
								"type": "string",
								"description": "Author username"
							},
							"profile_image_url": {
								"type": "string",
								"format": "uri",
								"description": "URL to author's profile image"
							}
						},
						"required": ["id", "username"],
						"description": "Post author information"
					},
					"likes_count": {
						"type": "integer",
						"description": "Number of likes on this post"
					},
					"comments_count": {
						"type": "integer",
						"description": "Number of comments on this post"
					},
					"is_liked": {
						"type": "boolean",
						"description": "Whether the authenticated user has liked this post. Can be false for newly created posts."
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"description": "Post creation timestamp"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"description": "Last post update timestamp"
					}
				},
				"required": [
					"id",
					"content",
					"author",
					"likes_count",
					"comments_count",
					"created_at"
				],
				"title": "PostDetail",
				"description": "Detailed post information"
			},
			"PostList": {
				"type": "object",
				"properties": {
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/PostDetail"
						},
						"description": "List of posts"
					},
					"total": {
						"type": "integer",
						"description": "Total number of posts"
					},
					"page": {
						"type": "integer",
						"description": "Current page number"
					},
					"size": {
						"type": "integer",
						"description": "Number of items per page"
					},
					"pages": {
						"type": "integer",
						"description": "Total number of pages"
					}
				},
				"required": ["items", "total", "page", "size", "pages"],
				"title": "PostList",
				"description": "Paginated list of posts"
			},
...
			"LikeResponse": {
				"type": "object",
				"properties": {
					"post_id": {
						"type": "integer",
						"description": "Post ID"
					},
					"is_liked": {
						"type": "boolean",
						"description": "Whether the post is now liked by the user"
					},
					"likes_count": {
						"type": "integer",
						"description": "Updated total number of likes on the post"
					}
				},
				"required": ["post_id", "is_liked", "likes_count"],
				"title": "LikeResponse",
				"description": "Like/unlike operation result"
			},
...
		}
	}
}